logic: Wait 30mins before refilling the crab
We cannot actually detect when the crab has deflated.  So instead of
trying to use a pressure limit for that, let's set a very long wait time
before the crab gets air again.

The absolute limit is about 1h, after which the crab has sagged very
noticeably.  Let's set the period to 30 minues for now.